pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-support-ops-env"
version = "0.2.0"
description = "Support operations environment for OpenEnv Round 1"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Codex" },
]
keywords = ["openenv", "benchmark", "support-ops", "agent-evals", "incident-response"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "License :: OSI Approved :: BSD License",
    "Intended Audience :: Developers",
]
dependencies = [
    "openenv-core[core]>=0.2.3",
    "fastapi>=0.115.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.8.0",
    "openai>=1.40.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "support_ops_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["support_ops_env", "support_ops_env.server"]
package-dir = { "support_ops_env" = ".", "support_ops_env.server" = "server" }
